+++ hc/front/forms.py
@@ -108,3 +108,7 @@ class AddSmsForm(forms.Form):
     error_css_class = "has-error"
     label = forms.CharField(max_length=100, required=False)
     value = forms.CharField(max_length=16, validators=[phone_validator])
+
+
+class ChannelNameForm(forms.Form):
+    name = forms.CharField(max_length=100, required=False)