+++ hc/front/tests/test_update_channel_name.py
@@ -1,36 +1,36 @@
-from hc.api.models import Check
+from hc.api.models import Channel
 from hc.test import BaseTestCase
 
 
-class UpdateNameTestCase(BaseTestCase):
+class UpdateChannelNameTestCase(BaseTestCase):
 
     def setUp(self):
-        super(UpdateNameTestCase, self).setUp()
-        self.check = Check(user=self.alice)
-        self.check.save()
+        super(UpdateChannelNameTestCase, self).setUp()
+        self.channel = Channel(kind="email", user=self.alice)
+        self.channel.save()
 
-        self.url = "/checks/%s/name/" % self.check.code
+        self.url = "/integrations/%s/name/" % self.channel.code
 
     def test_it_works(self):
-        payload = {"name": "Alice Was Here"}
+        payload = {"name": "My work email"}
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(self.url, data=payload)
-        self.assertRedirects(r, "/checks/")
+        self.assertRedirects(r, "/integrations/")
 
-        self.check.refresh_from_db()
-        self.assertEqual(self.check.name, "Alice Was Here")
+        self.channel.refresh_from_db()
+        self.assertEqual(self.channel.name, "My work email")
 
     def test_team_access_works(self):
-        payload = {"name": "Bob Was Here"}
+        payload = {"name": "Bob was here"}
 
         # Logging in as bob, not alice. Bob has team access so this
         # should work.
         self.client.login(username="bob@example.org", password="password")
         self.client.post(self.url, data=payload)
 
-        self.check.refresh_from_db()
-        self.assertEqual(self.check.name, "Bob Was Here")
+        self.channel.refresh_from_db()
+        self.assertEqual(self.channel.name, "Bob was here")
 
     def test_it_checks_ownership(self):
         payload = {"name": "Charlie Sent This"}
@@ -39,31 +39,14 @@ class UpdateNameTestCase(BaseTestCase):
         r = self.client.post(self.url, data=payload)
         self.assertEqual(r.status_code, 403)
 
-    def test_it_handles_bad_uuid(self):
-        url = "/checks/not-uuid/name/"
-        payload = {"name": "Alice Was Here"}
-
-        self.client.login(username="alice@example.org", password="password")
-        r = self.client.post(url, data=payload)
-        self.assertEqual(r.status_code, 404)
-
     def test_it_handles_missing_uuid(self):
         # Valid UUID but there is no check for it:
-        url = "/checks/6837d6ec-fc08-4da5-a67f-08a9ed1ccf62/name/"
+        url = "/integrations/6837d6ec-fc08-4da5-a67f-08a9ed1ccf62/name/"
         payload = {"name": "Alice Was Here"}
 
         self.client.login(username="alice@example.org", password="password")
         r = self.client.post(url, data=payload)
         self.assertEqual(r.status_code, 404)
-
-    def test_it_sanitizes_tags(self):
-        payload = {"tags": "  foo  bar\r\t \n  baz \n"}
-
-        self.client.login(username="alice@example.org", password="password")
-        self.client.post(self.url, data=payload)
-
-        self.check.refresh_from_db()
-        self.assertEqual(self.check.tags, "foo bar baz")
 
     def test_it_rejects_get(self):
         self.client.login(username="alice@example.org", password="password")