+++ hc/front/urls.py
@@ -38,6 +38,7 @@ channel_urls = [
     path('add_trello/', views.add_trello, name="hc-add-trello"),
     path('add_trello/settings/', views.trello_settings, name="hc-trello-settings"),
     path('<uuid:code>/checks/', views.channel_checks, name="hc-channel-checks"),
+    path('<uuid:code>/name/', views.update_channel_name, name="hc-channel-name"),
     path('<uuid:code>/remove/', views.remove_channel, name="hc-remove-channel"),
     path('<uuid:code>/verify/<slug:token>/', views.verify_email,
          name="hc-verify-email"),