+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to this project will be documented in this file.
 - Add "channels" attribute to the Check API resource
 - Can specify channel codes when updating a check via API
 - Added a workaround for email agents automatically opening "Unsubscribe" links
+- Add Channel.name field, users can now name integrations
 
 ### Bug Fixes
 - During DST transition, handle ambiguous dates as pre-transition